removed predict_pieces method and added .run(inputs) instead
