+++ backend/logic/machine-learning/game/game_store.py
@@ -1,4 +1,4 @@
-from game import Game
+from game.game import Game
 
 class GameStore:
     def __init__(self):
